requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.6
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
playwright-stealth>=1.0.6
